refactor: vo 패키지 생성
- Price 클래스 패키지 domain -> vo 변경
