+++ src/main/java/kitchenpos/dao/JdbcTemplateMenuDao.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 import javax.sql.DataSource;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Price;
+import kitchenpos.vo.Price;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
