+++ src/main/java/kitchenpos/dto/MenuRequest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Price;
+import kitchenpos.vo.Price;
 
 public class MenuRequest {
